Move `auction_models` to `schema`
